requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2 
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.6.2
fastapi==0.95.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
pytest==7.2.1
pytest-asyncio==0.20.3
httpx==0.23.3